refactor: extract rest store response handling and tidy ffi auth docs
- Move HTTP status mapping into rest_store/response.rs
- Rename AuthCollector to AuthHeaderEmission
- Clarify CAuthHeaderCollector emit helper documentation
- Fix unused Error import when default-engine-base is disabled
